DI: restructure, move Injector interface to separate file
