+++ lib/src/core/di/public_injector.dart
@@ -3,14 +3,15 @@
 // found in the LICENSE file.
 
 import 'injectable.dart';
-import 'injector.dart' as Private;
+import 'injector.dart' as Interface;
+import 'private_injector.dart' as Private;
 
 /// Facade for public API usage.
 ///
 /// Since retrieved [Injector] instances shall only be used to retrieve
 /// dependencies and not overwrite anything, the public [Injector] is used to
 /// cover that. The public [Injector] proxies a [Private.Injector].
-class Injector implements Private.InjectorInterface {
+class Injector implements Interface.Injector {
   final Private.Injector _injector;
 
   Injector._internal(this._injector) {
